Remove V2 namespace
