+++ app/forms/find/subjects/secondary_form.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Find
+  module Subjects
+    class SecondaryForm
+      include ActiveModel::Model
+
+      attr_accessor :subjects
+
+      validates :subjects, presence: true
+
+      def initialize(params = {})
+        @subjects = Array(params[:subjects]).compact_blank
+      end
+    end
+  end
+end
