--- app/forms/find/v2/subjects/secondary_form.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-module Find
-  module V2
-    module Subjects
-      class SecondaryForm
-        include ActiveModel::Model
-
-        attr_accessor :subjects
-
-        validates :subjects, presence: true
-
-        def initialize(params = {})
-          @subjects = Array(params[:subjects]).compact_blank
-        end
-      end
-    end
-  end
-end
