+++ app/helpers/find/financial_incentive_helper.rb
@@ -13,14 +13,14 @@ module Find
 
       content = if scholarship && bursary
                   I18n.t(
-                    '.find.v2.subjects.fee_value.fee.hint.bursaries_and_scholarship_html',
+                    '.find.subjects.fee_value.fee.hint.bursaries_and_scholarship_html',
                     bursary_amount: number_to_currency(bursary),
                     scholarship_amount: number_to_currency(scholarship)
                   )
                 elsif scholarship
-                  I18n.t('.find.v2.subjects.fee_value.fee.hint.scholarship_only_html', scholarship_amount: number_to_currency(scholarship))
+                  I18n.t('.find.subjects.fee_value.fee.hint.scholarship_only_html', scholarship_amount: number_to_currency(scholarship))
                 elsif bursary
-                  I18n.t('.find.v2.subjects.fee_value.fee.hint.bursaries_only_html', bursary_amount: number_to_currency(bursary))
+                  I18n.t('.find.subjects.fee_value.fee.hint.bursaries_only_html', bursary_amount: number_to_currency(bursary))
                 end
 
       content_tag(:p, content, class: 'govuk-hint govuk-!-font-size-16 govuk-!-margin-top-0 govuk-!-margin-bottom-0') if content
