+++ app/views/find/homepage/index.html.erb
@@ -51,7 +51,7 @@
             </summary>
 
             <div class="govuk-details__text">
-              <div class="govuk-form-group" data-controller="v2-provider-autocomplete">
+              <div class="govuk-form-group" data-controller="provider-autocomplete">
                 <%= render Find::AutocompleteComponent.new(
                   form_field: form.govuk_select(
                     :provider_code,
