+++ app/views/find/primary_subjects/index.erb
