+++ app/views/find/results/courses_guidance/_show.html.erb
