+++ app/views/find/results/filters/_all.html.erb
