+++ config/routes/find.rb
@@ -22,11 +22,11 @@ namespace :find, path: '/' do
 
   get '/geolocation-suggestions', to: 'geolocation_suggestions#index'
 
-  get '/results', to: 'v2/results#index', as: 'results'
-  get '/primary', to: 'v2/primary_subjects#index'
-  post '/primary', to: 'v2/primary_subjects#submit'
-  get '/secondary', to: 'v2/secondary_subjects#index'
-  post '/secondary', to: 'v2/secondary_subjects#submit'
+  get '/results', to: 'results#index', as: 'results'
+  get '/primary', to: 'primary_subjects#index'
+  post '/primary', to: 'primary_subjects#submit'
+  get '/secondary', to: 'secondary_subjects#index'
+  post '/secondary', to: 'secondary_subjects#submit'
 
   get '/cycle-has-ended', to: 'pages#cycle_has_ended', as: 'cycle_has_ended'
